Custom keybindings for basic code refactoring
CoC has a funciton called CocSearch, which searches for all occurances
of given text recursively reachable from the current directory. Two
keybindings have been created with this feature. The first is CoC
searching the current word under the cursor, and the second is for a
string within a visual selection. Both open a separate buffer with the
matches from all files contained. Lastly, there is :GFiles, which is a
funciton of FZF's plugin that uses vim-fugitive to fuzzy search files in a
git project.